+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Support for pattern matching syntax
+
 ## [0.15.1] - 2023-02-08
 
 ### Fixed